Project and seeders
Added seeder for Story
Created Project (model, seeder, controller) to begin additional work on #8

+++ database/migrations/2021_03_07_111811_create_stories_table.php
@@ -15,7 +15,7 @@ class CreateStoriesTable extends Migration
     {
         Schema::create('stories', function (Blueprint $table) {
             $table->id();
-            $table->integer('project_id')->nullable();
+            $table->unsignedBigInteger('project_id')->nullable();
             $table->string('title');
             $table->text('description');
             $table->text('tests');
